PATCH: 8.2 fixes

+++ src/Tasks/SaveAllData.php
@@ -9,11 +9,12 @@ use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\DB;
 use SilverStripe\Versioned\Versioned;
 use Page;
+use SilverStripe\Control\Director;
 
 /**
  * A task to manually flush InterventionBackend cache.
  */
-class SaveAll extends BuildTask
+class SaveAllData extends BuildTask
 {
     protected $title = 'Write all dataobjects - use with extreme caution.';
 
@@ -29,28 +30,43 @@ class SaveAll extends BuildTask
     {
         DataObject::config()->set('validation_enabled', false);
         $classes = ClassInfo::subclassesFor(DataObject::class, false);
+        if(!Director::isDev()) {
+            die('you can only run this in dev mode');
+        }
+        $this->writeTableHeader();
+
         foreach ($classes as $class) {
             $singleton = Injector::inst()->get($class);
-            echo '<h2>' . $singleton->i18n_singular_name() . '</h2>';
+            $type =  $singleton->i18n_singular_name() . ' - ' . $singleton->ClassName;
             if($singleton->canEdit()) {
                 $list = $class::get()->orderBy('RAND()')->limit(10);
+                $action = 'write';
                 foreach ($list as $obj) {
-                    DB::alteration_message('Writing ' . $obj->getTitle());
+                    $timeBefore = microtime(true);
+                    $title = (string) $obj->getTitle() ?: (string) $obj->ID;
                     if($obj->hasExtension(Versioned::class)) {
                         $isPublished = $obj->isPublished() && !$obj->isModifiedOnDraft();
                         $obj->writeToStage(Versioned::DRAFT);
                         if($isPublished) {
+                            $action .= ' and publish';
                             $obj->publishSingle();
                         }
                     } else {
                         $obj->write();
                     }
+                    $this->writeTableRow($action, $type, $title, $timeBefore);
                 }
             } else {
-                DB::alteration_message('No permission to edit ' . $singleton->i18n_singular_name());
+                $action = 'write (not allowed)';
+                $title = 'n/a';
+                $timeBefore = microtime(true);
+                $this->writeTableRow($action, $type, $title, $timeBefore);
             }
             $createdObj = null;
             if($singleton->canCreate()) {
+                $timeBefore = microtime(true);
+                $action = 'create';
+                $title = 'NEW OBJECT';
                 $createdObj = $class::create();
                 if($createdObj->hasExtension(Versioned::class)) {
                     $createdObj->writeToStage(Versioned::DRAFT);
@@ -58,22 +74,74 @@ class SaveAll extends BuildTask
                         $createdObj->publishSingle();
                     }
                 } else {
-                    $createdObj->write();
+                    try {
+                        //$createdObj->write();
+                    } catch (\Exception $e) {
+                        $action = 'create (failed)';
+                        $title = $e->getMessage();
+                    }
                 }
             } else {
-                DB::alteration_message('No permission to create ' . $singleton->i18n_singular_name());
+                $action = 'creatre (not allowed)';
+                $title = 'n/a';
+                $timeBefore = microtime(true);
             }
-            if($createdObj) {
+            $this->writeTableRow($action, $type, $title, $timeBefore);
+            if($createdObj && $createdObj->exists()) {
+                $action = 'delete';
+                $title = (string) $createdObj->getTitle() ?: (string) $createdObj->ID;
+                $timeBefore = microtime(true);
                 if($createdObj->hasExtension(Versioned::class)) {
                     $createdObj->doUnpublish();
                     $createdObj->delete();
+                    $action .= ' and unpublish';
                 } else {
-                    $createdObj->deleted();
+                    $createdObj->delete();
                 }
-                DB::alteration_message('Deleted new items of ' . $singleton->i18n_singular_name());
+            } else {
+                $action = 'delete (not required)';
+                $title = 'n/a';
+                $timeBefore = microtime(true);
             }
+            $this->writeTableRow($action, $type, $title, $timeBefore);
 
         }
+        $this->writeTableFooter();
         DB::alteration_message('-----------------DONE ------------------');
+    }
+
+    protected function writeTableHeader()
+    {
+        echo '
+            <table>
+                <thead>
+                    <tr>
+                        <th>Action</th>
+                        <th>Type</th>
+                        <th>Name</th>
+                        <th>Time Taken</th>
+                    </tr>
+                </thead>
+            <tbody>';
+    }
+
+    protected function writeTableFooter()
+    {
+        echo '
+            </tbody>
+            </teable>';
+    }
+
+    protected function writeTableRow(string $action, string $type, string $title, float $timeBefore)
+    {
+        $timeAfter = microtime(true);
+        $timeTaken = round($timeAfter - $timeBefore, 2);
+        echo '
+            <tr>
+                <td>' . $action . '</td>
+                <td>' . $type . '</td>
+                <td>' . $title . '</td>
+                <td>' . $timeTaken . '</td>
+            </tr>';
     }
 }
